successful loading of both dataset v2 and v3
