+++ lerobot/common/datasets/v3/pipeline_features.py
@@ -19,7 +19,7 @@ from typing import Any
 from lerobot.configs.types import PipelineFeatureType
 from .utils import hw_to_dataset_features
 from lerobot.processor import DataProcessorPipeline
-from lerobot.utils.constants import ACTION, OBS_IMAGES, OBS_STATE, OBS_STR
+from lerobot.common.utils.constants import ACTION, OBS_IMAGES, OBS_STATE, OBS_STR
 
 
 def create_initial_features(
