+++ lerobot/common/policies/__init__.py
@@ -12,4 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""
+Importing the configuration modules here ensures they are registered with
+`PreTrainedConfig.register_subclass` as soon as the policies package is used.
+"""
+
+from .act.configuration_act import ACTConfig as ACTConfig
+from .diffusion.configuration_diffusion import DiffusionConfig as DiffusionConfig
 from .pi0.configuration_pi0 import PI0Config as PI0Config
+from .pi0fast.configuration_pi0fast import PI0FASTConfig as PI0FASTConfig
+from .tdmpc.configuration_tdmpc import TDMPCConfig as TDMPCConfig
+from .vqbet.configuration_vqbet import VQBeTConfig as VQBeTConfig
+
+__all__ = [
+    "ACTConfig",
+    "DiffusionConfig",
+    "PI0Config",
+    "PI0FASTConfig",
+    "TDMPCConfig",
+    "VQBeTConfig",
+]
